FIX: Cast std to float32 before weight initialization
ROOT CAUSE: In _init_weight(), the order was:
  np.random.randn(*shape).astype("float32") * std

This converted to float32, then multiplied by float64 std,
promoting the entire result BACK to float64!

SOLUTION: Cast std to float32 first, do multiplication in float64,
then cast the final result to float32:
  (np.random.randn(*shape) * np.float32(std)).astype("float32")

+++ examples/onnx/onnx-yolo-demo/blocks.py
@@ -90,8 +90,8 @@ class ConvBNSiLU:
     def _init_weight(self, shape, name):
         """He initialization."""
         fan_in = shape[1] * shape[2] * shape[3]  # in_channels * kh * kw
-        std = np.sqrt(2.0 / fan_in)
-        W_val = np.random.randn(*shape).astype("float32") * std
+        std = np.float32(np.sqrt(2.0 / fan_in))
+        W_val = (np.random.randn(*shape) * std).astype("float32")
         return shared(W_val, name=name, borrow=True)
 
     def __call__(self, x):
